Adicionando metodo para conversao de valor brl

+++ src/metricas.py
@@ -21,14 +21,14 @@ def indicadores_total_mes(balanco):
     
     #total despesa variaveis
     df_depesa_variavel_atual = df_despesas[df_despesas["balanco"] == balanco]
-    total_despesa_variavel = int(df_depesa_variavel_atual["valor"].sum())
+    total_despesa_variavel = float(df_depesa_variavel_atual["valor"].sum())
 
     #total despesas fixas
-    total_despesa_fixa = int(df_despesas_fixas["valor"].sum())
+    total_despesa_fixa = float(df_despesas_fixas["valor"].sum())
 
     #total ganho
     df_ganho_atual = df_receita[df_receita["balanco"] == balanco]
-    total_ganho = int(df_ganho_atual["valor"].sum())
+    total_ganho = float(df_ganho_atual["valor"].sum())
 
     #Sobra
     tota_sobra = total_ganho - (total_despesa_variavel + total_despesa_fixa)
